refactor: Dismember huge and messy background.js
