+++ app/scripts/constants.js
@@ -2,9 +2,11 @@
 // S -> Service
 export default {
   MESSAGE_ROUTER: Symbol('router'),
-  R_ACCESS_TOKEN: Symbol('accessTokenRepository'),
-  S_CACHE: Symbol('cacheService'),
-  S_CHROME_STORAGE: Symbol('chromeStorageService'),
-  S_COMMON: Symbol('commonService'),
-  S_GITHUB: Symbol('githubService'),
+  R_ACCESS_TOKEN: Symbol('access token repository'),
+  S_CACHE: Symbol('cache service'),
+  S_CHROME_STORAGE: Symbol('chrome storage service'),
+  S_COMMON: Symbol('common service'),
+  S_GITHUB: Symbol('github service'),
+  LOG: Symbol('log function'),
+  UPDATE_BADGE: Symbol('update badge function'),
 };
