+++ package.json
@@ -46,7 +46,7 @@
     "gulp-cache": "0.x.x",
     "gulp-clean-css": "3.x.x",
     "gulp-filter": "4.x.x",
-    "gulp-git": "1.x.x",
+    "gulp-git": "2.x.x",
     "gulp-if": "2.x.x",
     "gulp-imagemin": "3.x.x",
     "gulp-json-transform": "0.x.x",
